Dynamic inference example | Remove shorten_str().

See merge request ADLR/megatron-lm!3960

+++ examples/inference/gpt/utils.py
@@ -226,7 +226,7 @@ def get_synthetic_requests(args: Namespace, tokenizer: Any) -> list[Request]:
         args.seed,
         args.incoming_requests_per_step,
         args.incoming_requests_per_sec,
-        args.incoming_requests_per_sec * args.incoming_requests_duration,
+        int(args.incoming_requests_per_sec * args.incoming_requests_duration),
     )
 
     # Init requests.
